+++ src/_P058_HT16K33_KeyPad.ino
@@ -76,7 +76,7 @@ boolean Plugin_058(uint8_t function, struct EventStruct *event, String& string)
     case PLUGIN_I2C_HAS_ADDRESS:
     case PLUGIN_WEBFORM_SHOW_I2C_PARAMS:
     {
-      const int i2cAddressValues[] = { 0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77 };
+      const uint8_t i2cAddressValues[] = { 0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77 };
       if (function == PLUGIN_WEBFORM_SHOW_I2C_PARAMS) {
         addFormSelectorI2C(F("i2c_addr"), 8, i2cAddressValues, PCONFIG(0));
       } else {
@@ -123,7 +123,7 @@ boolean Plugin_058(uint8_t function, struct EventStruct *event, String& string)
 
         if (P058_data->readKey(key))
         {
-          UserVar[event->BaseVarIndex] = (float)key;
+          UserVar[event->BaseVarIndex] = key;
           event->sensorType            = Sensor_VType::SENSOR_TYPE_SWITCH;
 
           if (loglevelActiveFor(LOG_LEVEL_INFO)) {
